requirements.txt
discord.py==2.5.1
python-dotenv>=0.21.0
